Updated gas-fakes.js
